clojure-clojurescript-buck: ClojureSctipt test support
cljs_test.sh file added which extracts all the namespaces from test files and create a test runner which depends on those tests. Later on test runner loads and execute all the tests. Right now NS extractring is done using magic of sed and regular expression. In a feature we can (should/have) to replace it with proper CLJ reader