RULES/clojure-clojurescript-buck: Tests runs 3 times faster now
Previously before running tests we were building it first separetly, which took extra time (around ~30s for CLJS and ~8s for CLJ).
We we running tests two times: For debug and for advanced build profiles which requires one more extra build (~40s)
We were using :simple for CLJS optimization level which required additional 5s comparing to simplest :whitespace

Now we have native speed compared to simple lein, no extra work is done.

Eventually we should create `clj*-release` task which would be running test with advanced optimization but only for end project and not for each of the libs