RULES/clojure-clojurescript-buck: Test deps and REPL implemented
As discussed in https://github.com/artemyarulin/clojure-clojurescript-buck/issues/22 we eliminated any additional logic for testing - now test is just a module with it's own source files which depends on some other modules plus a flag that test runner has to be created. Doing that we were able to simplify lib.py and shift some of the logic to config.py. Closes https://github.com/artemyarulin/clojure-clojurescript-buck/issues/21

Initial implementation of REPL - again more details could be found in #22. It works fine, it's not yet ready to be released as sync is very slow and certain things (like :ext) are still hardcoded in the source, while it should be configurable. Closes https://github.com/artemyarulin/clojure-clojurescript-buck/issues/19